Issue 109: Move createName to other file
